update for full release
First update here in a looooong time, so this includes lots and lots of changes. It still includes unused code for creating the xml for model cards that needs to be cleaned up. I was trying to see if it would speed up loading if the server handled the xml formatting rather than tts, but strangely it didn't make much of a difference, so I left it in tts to save on data being transferred to and from the server.